Okay, so finally: Latch is done!!!! Here's what I did: Added the double solenoid, added timer so that if the latch is opened for more than a second it will close. All of this is in Latch.h and Latch.cpp, I deleted all the other unnecessary files.
