+++ src/RobotMap.h
@@ -46,8 +46,8 @@ const uint32_t ELEVATOR_BOTTOM_SWITCH = 1;
 const uint32_t ELEVATOR_ENCODER_A = 1;
 const uint32_t ELEVATOR_ENCODER_B = 1;
 
-const uint32_t FIRST_SOLENOID = 0;
-const uint32_t SECOND_SOLENOID = 1;
+const uint32_t FIRST_SOLENOID = 0; // check correct values
+const uint32_t SECOND_SOLENOID = 1; // check correct values
 
 const float MOTOR_EXPIRATION = 0.1f; // float measured in seconds
