+++ src/Subsystems/DriveAccelerometer.h
@@ -1,5 +1,5 @@
-#ifndef SRC_SUBSYSTEMS_DRIVEACCELEROMETER_H_
-#define SRC_SUBSYSTEMS_DRIVEACCELEROMETER_H_
+#ifndef LATCH_CPP_SRC_SUBSYSTEMS_DRIVEACCELEROMETER_H_
+#define LATCH_CPP_SRC_SUBSYSTEMS_DRIVEACCELEROMETER_H_
 
 #include "Commands/Subsystem.h"
 #include "WPILib.h"
@@ -19,4 +19,4 @@ public:
 	double GetZ();
 };
 
-#endif /* SRC_SUBSYSTEMS_DRIVEACCELEROMETER_H_ */
+#endif /* LATCH_CPP_SRC_SUBSYSTEMS_DRIVEACCELEROMETER_H_ */
