+++ src/Subsystems/Elevator.cpp
@@ -7,8 +7,6 @@ Elevator::Elevator() :
 	topSwitch = new DigitalInput(ELEVATOR_TOP_SWITCH);
 	bottomSwitch = new DigitalInput(ELEVATOR_BOTTOM_SWITCH);
 	encoder = new Encoder(ELEVATOR_ENCODER_A, ELEVATOR_ENCODER_B);
-	firstsolenoid = new DoubleSolenoid(FIRST_SOLENOID);
-	secondsolenoid = new DoubleSolenoid(SECOND_SOLENOID);
 }
 
 Elevator::~Elevator()
