+++ src/Subsystems/Elevator.h
@@ -1,5 +1,5 @@
-#ifndef SRC_SUBSYSTEMS_ELEVATOR_H_
-#define SRC_SUBSYSTEMS_ELEVATOR_H_
+#ifndef LATCH_CPP_SRC_SUBSYSTEMS_ELEVATOR_H_
+#define LATCH_CPP_SRC_SUBSYSTEMS_ELEVATOR_H_
 
 #include "Commands/Subsystem.h"
 #include "WPILib.h"
@@ -28,4 +28,4 @@ public:
 	void secondSolOpen();
 };
 
-#endif /* SRC_SUBSYSTEMS_ELEVATOR_H_ */
+#endif /* LATCH_CPP_SRC_SUBSYSTEMS_ELEVATOR_H_ */
